+++ packages/web-component-library-vue/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @utrecht/web-component-library-vue
 
+## 3.0.3
+
+### Patch Changes
+
+- Updated dependencies [ffc9357]
+  - @utrecht/web-component-library-stencil@4.2.0
+
 ## 3.0.2
 
 ### Patch Changes